services/alarm-processor/requirements.txt
psycopg2-binary==2.9.9
jsonpath-ng==1.5.3
kafka-python==2.0.2
